Added structure validation
